MAPPS-32 Fix crash for 73 and 96
An invalid PolyStyle element was included in the KML files for 73 and 96. This was causing the app to crash because we were force unwrapping the element.

Removed the invalid XML from the two KML files and changed the code to use default values rather than force unwrap so that future invalid KML won’t cause a crash.