Fix bug where favorites were not showing up
We had added two new properties to Favorite: collapsed, and sortOrder. For any users who previously had saved favorites, and then update to this new version, would have their favorites no longer appear. The reason was when we were decoding the objects, we were throwing DecodingError.keyNotFound (because it's a new property that didn't exist when the Favorite was encoded) causing the object to not successfully initialize. Adding the catch block gives us the change to set a default value.